Fix static array that was needlessly declared double.

+++ vtkDICOMGenerator.cxx
@@ -167,7 +167,7 @@ void vtkDICOMGenerator::ComputeAspectRatio(
   const double spacing[2], int aspect[2])
 {
   // compute aspect ratio to within one part in 1000
-  static const double primes[] = { // include 1
+  static const short primes[] = { // include 1
     1, 2, 3, 5, 7, 11, 13, 17, 19, 23, 29,
     31, 37, 41, 43, 47, 53, 59, 61, 67, 71,
     73, 79, 83, 89, 97, 101, 103, 107, 109, 113,
